+++ spotify_mockup/lib/settings.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class Settings extends StatelessWidget {
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Settings'),
+    );
+  }
+}
