+++ spotify_mockup/pubspec.yaml
@@ -29,7 +29,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_launcher_icons:
-  spotify_playback:
+#  spotify_playback:
 flutter_icons:
   image_path: assets/spotify-2015-logo-560E071CB7-seeklogo.com.png
   android: true
